feat: Refactor syllabi handling by consolidating SyllabiService into WebsocService and updating related schemas
