+++ apps/api/src/schema/index.ts
@@ -14,5 +14,4 @@ export * from "./programs";
 export * from "./search";
 export * from "./study-rooms";
 export * from "./websoc";
-export * from "./websoc-syllabi";
 export * from "./week";
