+++ apps/api/src/services/index.ts
@@ -11,5 +11,4 @@ export * from "./programs";
 export * from "./search";
 export * from "./study-rooms";
 export * from "./websoc";
-export * from "./websoc-syllabi";
 export * from "./week";
